Added components folder
