+++ frontend/src/components/pictures/Digital-image.png
