+++ frontend/src/components/pictures/gradient-image.jpg
