+++ frontend/src/components/pictures/white-section.png
